Update _byo-ca-mtls-cert-requirements.md
I am adding this line to make clear customer that a leaf cert will be not valid. I had few tickets with this issue. 

We are already telling this on the API with this error :

"code": 1470,"message": "CA certificate chain contains a leaf certificate."

However, customer are getting confused. If we can add this to the public documentation, they cannot discuss this further. 

I could also consider adding some explanation such as:

A root certificate is required to be able to issue certificates on behalf the CA issuing the certificate. A leaf certificate will be no valid. 

Example of a CUSTESC: https://jira.cfdata.org/browse/CUSTESC-33051